+++ src/main/java/com/rocket/user/userfeed/service/FeedService.java
@@ -2,16 +2,21 @@ package com.rocket.user.userfeed.service;
 
 import static com.rocket.error.type.UserErrorCode.USER_DELETED_AT;
 import static com.rocket.error.type.UserErrorCode.USER_NOT_FOUND;
+import static com.rocket.error.type.UserFeedErrorCode.FEED_IMAGE_UPLOAD_COUNT_OVER;
 import static com.rocket.error.type.UserFeedErrorCode.FEED_NOT_FOUND;
+import static com.rocket.error.type.UserFeedErrorCode.FEED_UPDATE_FAIL;
+import static com.rocket.error.type.UserFeedErrorCode.FEED_USER_NOT_MATCH;
 
 import com.rocket.error.exception.UserException;
 import com.rocket.error.exception.UserFeedException;
+import com.rocket.error.type.UserFeedErrorCode;
 import com.rocket.user.user.entity.User;
 import com.rocket.user.user.repository.UserRepository;
 import com.rocket.user.userfeed.dto.FeedDto;
 import com.rocket.user.userfeed.dto.FeedSearchCondition;
 import com.rocket.user.userfeed.entity.Feed;
 import com.rocket.user.userfeed.repository.FeedRepository;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
@@ -20,6 +25,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 @Slf4j
 @Service
@@ -28,6 +34,7 @@ public class FeedService {
 
     private final FeedRepository feedRepository;
     private final UserRepository userRepository;
+    private final FeedImageService feedImageService;
 
     @Transactional(readOnly = true)
     public User getUser(String uuid) {
@@ -41,10 +48,16 @@ public class FeedService {
     }
 
     public Feed getFeed(Long feedId) {
-        return feedRepository.findById(feedId).orElse(null);
+
+        Feed getFeed = feedRepository.findById(feedId).orElse(null);
+
+        if (getFeed == null) {
+            throw new UserFeedException(UserFeedErrorCode.FEED_NOT_FOUND);
+        }
+        return getFeed;
     }
 
-    public Page<Feed> getFeeds(User user, FeedSearchCondition searchCondition,
+    public Page<Feed> getFeedList(User user, FeedSearchCondition searchCondition,
         Pageable pageable) {
 
         return feedRepository.findByUserIdAndRcate1EqualsAndRcate2EqualsOrderByCreatedAtDesc(
@@ -54,22 +67,44 @@ public class FeedService {
             , PageRequest.of(pageable.getPageNumber(), pageable.getPageSize()));
     }
 
+    public Page<Feed> getFeeds(FeedSearchCondition searchCondition,
+        Pageable pageable) {
+
+        return feedRepository.findByRcate1EqualsAndRcate2EqualsOrderByCreatedAtDesc(
+            searchCondition.getRcate1()
+            , searchCondition.getRcate2()
+            , PageRequest.of(pageable.getPageNumber(), pageable.getPageSize()));
+    }
+
     @Transactional
-    public FeedDto createFeed(User user, FeedDto feedDto) {
+    public FeedDto createFeed(User user, Feed feed, List<MultipartFile> multipartFiles) {
 
-        Feed feed = feedRepository.save(Feed.builder()
-            .user(user)
-            .title(feedDto.getTitle())
-            .content(feedDto.getContent())
-            .rcate1(feedDto.getRcate1())
-            .rcate2(feedDto.getRcate2())
-            .rcate3(feedDto.getRcate3())
-            .longitude(feedDto.getLongitude())
-            .latitude(feedDto.getLatitude())
-            .build());
-        log.info("feed {}", feed);
+        Feed newFeed = null;
 
-        return new ModelMapper().map(feed, FeedDto.class);
+        if (multipartFiles.size() > 4) {
+            throw new UserFeedException(FEED_IMAGE_UPLOAD_COUNT_OVER);
+        }
+
+        try {
+            newFeed = feedRepository.save(Feed.builder()
+                .user(user)
+                .title(feed.getTitle())
+                .content(feed.getContent())
+                .rcate1(feed.getRcate1())
+                .rcate2(feed.getRcate2())
+                .longitude(feed.getLongitude())
+                .latitude(feed.getLatitude())
+                .build());
+
+            feedImageService.createFeedImage(user, newFeed, multipartFiles);
+
+        } catch (Exception e) {
+            log.error("[FeedService.createFeed] ERROR {}", e);
+
+            throw new UserFeedException(UserFeedErrorCode.FEED_CREATE_FAIL);
+        }
+        return new ModelMapper().map(newFeed, FeedDto.class);
+
     }
 
     @Transactional
@@ -77,24 +112,27 @@ public class FeedService {
 
         Feed feed = feedRepository.findById(id)
             .orElseThrow(() -> new UserFeedException(FEED_NOT_FOUND));
-        feed.updateFeed(feedDto);
+
+        try {
+            feed.updateFeed(feedDto);
+        } catch (Exception e) {
+            log.error("[FeedService.updateFeed] ERROR {}", e.getMessage());
+            throw new UserFeedException(FEED_UPDATE_FAIL);
+        }
 
         return new ModelMapper().map(feed, FeedDto.class);
     }
 
     @Transactional
-    public void deleteFeed(Long userId, Long FeedId) {
+    public void deleteFeed(Long userId, Long feedId) {
 
-        Feed feed = feedRepository.findById(FeedId)
+        Feed feed = feedRepository.findById(feedId)
             .orElseThrow(() -> new UserFeedException(FEED_NOT_FOUND));
 
-        try {
+        if (feed.getUser().getId().equals(userId)) {
             feedRepository.delete(feed);
-        } catch (Exception e) {
-            log.error("[FeedService.deleteFeed] ERROR {}", e.getMessage());
-
-            // TODO: 먼가 만들자
-            throw new UserFeedException();
+        } else {
+            throw new UserFeedException(FEED_USER_NOT_MATCH);
         }
     }
 }
